Move mock WeatherProvider to weather.spec.ts
